refactor utils directory into modules. Note for schema scaffold generators to do same \n close #27
