+++ src/utils/generators/color.js
@@ -1,0 +1,3 @@
+export default function color () {
+  return '#' + Math.floor(Math.random() * Math.pow(2, 24)).toString(16);
+}
